Refactor TextureViewDescription creation process
Removed the constructor from `TextureViewDescription` that took a `Texture` parameter. Added a static `Create` method to `TextureViewDescription` for creating instances with default `baseMipLevel` and `mipLevels` values. Updated `Program.cs` to use the new `Create` method for creating `TextureViewDescription` instances. Introduced an abstract `CreateTextureView` method in `ResourceFactory` and implemented it in `VKResourceFactory` to create `VKTextureView` instances.