Add Size property to LayoutElementDesc and update VKResourceSet
The LayoutElementDesc struct in LayoutElementDesc.cs now includes a new
property `Size` of type `uint`. This property, if greater than 0,
overrides the size of the resource (in bytes) and is only valid for
Buffers. The Default method of LayoutElementDesc has been modified to
accept an additional parameter `size` with a default value of 0, which
is used to set the new `Size` property.

In VKResourceSet.cs, a new using directive for Graphics.Engine.Enums has
been added. The VKResourceSet class has been updated to handle the new
Size property in LayoutElementDesc. When writing descriptor sets, it
now checks the Size property of each LayoutElementDesc and uses it to
set the Range of DescriptorBufferInfo if it is greater than 0. If the
Size is 0, it defaults to Vk.WholeSize.

The code now includes logic to handle different types of resources
(VKBuffer, VKTextureView, VKSampler) and appropriately sets the
PBufferInfo or PImageInfo pointers in the WriteDescriptorSet structure.
The UpdateDescriptorSets method is called to update the descriptor sets
with the new write information.