Add QueryHeapDesc structure and QueryHeapType enum
Introduce `QueryHeapDesc` in `ZenithEngine.Common.Descriptions`
to describe query heap properties with fields for type and count.
Add `QueryHeapType` enum in `ZenithEngine.Common.Enums`
with values `Occlusion` and `Timestamp`. Update the `QueryHeap`
class in `ZenithEngine.Common.Graphics` to utilize the new
`QueryHeapDesc` structure, including a constructor and a
public property for accessing the descriptor.